refactor: reformat & reorganize code
