+++ src/main/java/fr/moussax/blightedMC/BlightedMC.java
@@ -1,6 +1,7 @@
 package fr.moussax.blightedMC;
 
 import fr.moussax.blightedMC.commands.CommandBuilder;
+import fr.moussax.blightedMC.core.entities.listeners.BlightedEntitiesListener;
 import fr.moussax.blightedMC.core.registry.RegistrySystem;
 import org.bukkit.plugin.java.JavaPlugin;
 
@@ -13,12 +14,20 @@ public final class BlightedMC extends JavaPlugin {
     instance = this;
 
     RegistrySystem.initializeAllRegistries();
-
     eventsRegistry = new EventsRegistry();
     eventsRegistry.initializeListeners();
 
     CommandBuilder.initialize(this);
     CommandBuilder.initializeCommands();
+
+    // Rehydrate any loaded entities in already-loaded chunks (e.g., after /reload)
+    // by manually firing the chunk scan once on enabling.
+    // This complements the runtime ChunkLoadEvent handler.
+    getServer().getWorlds().forEach(world -> {
+      for (org.bukkit.Chunk chunk : world.getLoadedChunks()) {
+        BlightedEntitiesListener.rehydrateChunk(chunk);
+      }
+    });
   }
 
   @Override
@@ -27,7 +36,6 @@ public final class BlightedMC extends JavaPlugin {
     if (eventsRegistry != null && eventsRegistry.getBlockListener() != null) {
       eventsRegistry.getBlockListener().saveData();
     }
-
     RegistrySystem.clearAllRegistries();
   }
